refactor(compositor): handle cursor rendering and surface overflow
- Move cursor rendering into the compositor's back-buffer pass,
  removing the separate cursor-only render from the event handler.
  The compositor now uses prepare_frame() dirty rects to cover
  both the old and new cursor positions in one draw.

- Fix surface clipping to use window dimensions instead of source
  dimensions, and fill areas beyond the surface with a fallback
  background color. This properly supports windows that are larger
  than their surfaces (e.g., after tiling).

- Make update_taskbar() return a bool indicating whether the entry
  list changed (count or order).

- Remove a redundant i32 cast in window manager tiling code.

+++ lattice/src/wm.rs
@@ -541,7 +541,7 @@ impl WindowManager {
                 let stack_y = top + ((i as u32) * (each_h + gap)) as i32;
                 if let Some(sw) = self.windows.iter_mut().find(|w| w.id == id) {
                     sw.x = stack_left;
-                    sw.y = stack_y as i32;
+                    sw.y = stack_y;
                     sw.width = stack_w;
                     sw.height = each_h;
                     self.dirty_rects.push(window_dirty_rect(sw));
